Restore v8 archive flashing compatibility.
A field in the archive switched from a T to an Option<T> in the v9
format change. Unfortunately, this means old v8 archives are missing a
Some() that RON expects. My intention was to gloss over this issue by
enabling RON's IMPLICIT_SOME option... but I failed to.

It turns out that the flash loading stuff is not covered by Humility's
test suites, which is good to know.

This change enables IMPLICIT_SOME and works with both v8 and v9.