Allow building without usb/debug probes
Humility was originally designed to be used with a variety of
debug probes. It's since grown the ability to take dumps over the
network. There are situations/environments where it's useful to
run in network only mode. Allow building without debug probes
with `cargo build --no-default-features`.

To make this work with cargo features, the probes are pulled into
a separate crate and the humility binary is moved to `humilty-bin`