Fix cargo run and install.
Since xtask was added a while back, both cargo run and cargo install
have been broken, requiring you to specify a package.

This fixes that.

I've also forced the Clippy checks to apply the MSRV only, because
otherwise I happen to be the first commit after 1.90 was released with a
bunch of new Clippy warnings, which are preventing me from merging this.

And then I've changed the MSRV, because apparently the code wasn't
Clippy-clean at its MSRV. Wheeeee