Respect probe selection in `humility debugmailbox`
To access the debugmailbox, we have to jump through some extra
hoops to get the raw debug probe. We should still respect the
probe selected via the command line or `HUMILITY_PROBE` environment
variable. If no probe is specified we go with the existing behavior
of chosing the first probe, otherwise the probe must be specified
in `vid:pid:serial` format.