Use measurement token in `humility reset` (#580)
This is a little tricky, because we need a `ProbeCore` and Hubris archive to do
the reset-with-handoff logic.  If we don't have a valid archive, then we fall
back to the previous behavior of a plain reset - unless the user specifically
demands that we use the handoff token, in which case we return an error.