Fix LPC55 resetting from ISP mode (#583)
LPC55 must use `attach_to_probe` for initial programming.  In general,
`attach_to_probe` is more robust, so let's use it unless the caller (or archive)
specifically needs the measurement token handoff.

Tested on a benchtop, with
```
humility -a ../hubris/target/oxide-rot-1-selfsigned/dist/a/build-oxide-rot-1-selfsigned-image-a.zip debugmailbox isp
humility -a ../hubris/target/oxide-rot-1-selfsigned/dist/a/build-oxide-rot-1-selfsigned-image-a.zip reset
```

This fails on `master` but works with this patch.