rendmp: stop hardcoding enum-string conversion
This is preparing for adding more variants to SupportedDevices. The list
of SupportedDevices was present, explicitly or implicitly, in several
places in the code:

- The enum definition itself
- A list of strings used to filter device names
- A set of match statements used to convert those strings into variants
  of the enum
- Error messages listing support device types, which hardcoded the
  variant names.
- A Display impl that repeated the names.

This was noisy and potentially error-prone. I have simplified it by
applying the `strum` crate to derive traits for the SupportedDevice
macro:

- `FromStr` (derived by EnumString) eliminates all handrolled parsing
  and match code
- `VariantNames` (derived by EnumVariantNames) allows the error messages
  to list expected names without hardcoding them
- `Display` can now be derived automatically.

I chose `strum` because of its option to make string comparisons
case-insensitive, which neatly covers this command's use case of
converting lowercase manifest device names into uppercase enum variants.